requirements.txt
openai
llama-index
python-dotenv
PyMuPDF
tqdm
gradio
qdrant-client
colorlog
PyMuPDF
python-docx
